+++ CHANGELOG.md
@@ -21,5 +21,7 @@
   - [Work in progress] Define the `ClientManager` class to manage references to
     connected clients.
 - [Work in progress] Define the `Socket` class as the super-type of sockets.
+  - [Work in progress] Define the `HeartbeatManager` class to manage the
+    heartbeat mechanism.
   - [Work in progress] Define the server/`Socket` class to represent clients
     connected to the server.